Make system more modular, organize files
Moved core game logic to core
Moved responder logic to responders

Add ollama responder
