+++ basilisk/__init__.py
@@ -1,0 +1,6 @@
+# basilisk/__init__.py
+from core.game_manager import GameManager
+from core.game_state import SecretHitler
+from responders import HumanResponder, OllamaResponder, Responder
+
+__all__ = ['GameManager', 'SecretHitler', 'HumanResponder', 'OllamaResponder', 'Responder']
