+++ basilisk/core/__init__.py
@@ -1,0 +1,6 @@
+# core/__init__.py
+from .game_manager import GameManager
+from .game_state import SecretHitler
+from .input_handler import InputHandler
+
+__all__ = ['GameManager', 'SecretHitler', 'InputHandler']
