+++ basilisk/responders/human/__init__.py
@@ -1,0 +1,3 @@
+from .responder import HumanResponder
+
+__all__ = ['HumanResponder']
