+++ basilisk/responders/ollama/__init__,py
@@ -1,0 +1,4 @@
+# responders/ollama/__init__.py
+from .responder import OllamaResponder
+
+__all__ = ['OllamaResponder']
