+++ basilisk/responders/ollama/responder.py
@@ -1,0 +1,112 @@
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+import json
+import re
+from langchain_community.llms import Ollama
+
+from ..base import Responder, InputRequest
+
+@dataclass
+class Memory:
+    """Simple memory buffer for storing past events/decisions"""
+    max_items: int = 10
+    entries: List[str] = field(default_factory=list)
+    
+    def add(self, entry: str):
+        self.entries.append(entry)
+        if len(self.entries) > self.max_items:
+            self.entries.pop(0)
+    
+    def get_recent(self, limit: Optional[int] = None) -> str:
+        entries = self.entries[-limit:] if limit else self.entries
+        return "\n".join(entries) if entries else "No previous memories."
+
+class OllamaResponder(Responder):
+    def __init__(self, 
+                 system_prompt_path: Optional[str] = None,
+                 model_name: str = "llama2",
+                 memory_size: int = 10):
+        self.llm = Ollama(base_url="http://localhost:11434", model=model_name)
+        self.memory = Memory(max_items=memory_size)
+        
+        # Load system prompt
+        self.system_prompt = ""
+        if system_prompt_path:
+            path = Path(system_prompt_path)
+            if path.exists():
+                self.system_prompt = path.read_text()
+    
+    def _build_prompt(self, request: InputRequest) -> str:
+        """Build prompt with system instructions, memory, context, and required format"""
+        sections = []
+        
+        # Add system prompt if exists
+        if self.system_prompt:
+            sections.append(f"System Instructions:\n{self.system_prompt}\n")
+            
+        # Add memory context
+        sections.append(f"Recent Events:\n{self.memory.get_recent(5)}\n")
+        
+        # Add current context
+        sections.append(f"Current Situation:\n{request.context}\n")
+        
+        # Add required response format
+        sections.append(
+            "Please provide your response in the following format:\n\n"
+            "<memory_update>\n"
+            "Summarize the current situation and any important observations\n"
+            "</memory_update>\n\n"
+            
+            "<reasoning>\n"
+            "Explain your strategic thinking and analysis\n"
+            "</reasoning>\n\n"
+            
+            "<decision>\n"
+            f"Provide your decision in this JSON format:\n{json.dumps(request.example.values, indent=2)}\n"
+            "</decision>"
+        )
+        
+        return "\n".join(sections)
+    
+    def _extract_section(self, content: str, section: str) -> Optional[str]:
+        """Extract content from a section"""
+        pattern = f"<{section}>(.*?)</{section}>"
+        if match := re.search(pattern, content, re.DOTALL):
+            return match.group(1).strip()
+        return None
+    
+    def _parse_decision(self, decision_str: str) -> Dict[str, Any]:
+        """Parse and validate decision JSON"""
+        try:
+            return json.loads(decision_str)
+        except json.JSONDecodeError:
+            raise ValueError("Invalid decision JSON format")
+    
+    def get_response(self, request: InputRequest) -> Dict[str, Any]:
+        """Get response from model with retries"""
+        max_retries = 3
+        
+        for attempt in range(max_retries):
+            try:
+                # Get model response
+                prompt = self._build_prompt(request)
+                response = self.llm.generate([prompt])[0].text
+                
+                # Extract sections
+                if memory_update := self._extract_section(response, "memory_update"):
+                    self.memory.add(memory_update)
+                
+                if decision_str := self._extract_section(response, "decision"):
+                    decision = self._parse_decision(decision_str)
+                    
+                    # Validate required fields
+                    if all(field.name in decision for field in request.fields if field.required):
+                        return decision
+                    
+                raise ValueError("Missing required fields in decision")
+                
+            except Exception as e:
+                if attempt == max_retries - 1:
+                    raise ValueError(f"Failed to get valid response after {max_retries} attempts: {str(e)}")
+                continue
